+++ internal/kitty/run.go
@@ -9,6 +9,7 @@ import (
 const (
 	ListOSWindowsCmd      = "list-os-windows"
 	GotoOSWindowCmd       = "goto-os-window"
+	TargetsCmd            = "targets"
 	NewSessionCmd         = "new-session"
 	SessionsCmd           = "sessions"
 	DeleteSessionCmd      = "delete-session"